+++ projProblox/Source/projProblox/Collector.cpp
@@ -51,7 +51,7 @@ int ACollector::GetLvlCellCount(const ELevel& lvl) const
 	const UWorld* wrld = GetWorld();
 	TArray<AActor*> countArr;
 
-	unsigned int count = 0;
+	unsigned short count = 0;
 	
 	UGameplayStatics::GetAllActorsOfClass(wrld, ACellSpawner::StaticClass(), countArr);
 	for (const auto& spawnActor: countArr)
@@ -77,7 +77,7 @@ int ACollector::GetCollectedCountFromLvl(const ELevel& lvl) const
 	const UWorld* wrld = GetWorld();
 	TArray<AActor*> countArr;
 
-	unsigned int count = 0;
+	unsigned short count = 0;
 
 	UGameplayStatics::GetAllActorsOfClass(wrld, ACellSpawner::StaticClass(), countArr);
