+++ projProblox/Source/projProblox/Pickupables/BuyableAttachment.cpp
@@ -66,7 +66,7 @@ void ABuyableAttachment::SetHide(const bool hide) const
 	if(!hide)
 	{
 		const FBuyableInfoStruct buyInfo = info->GetInfo();
-		for(unsigned short i = 0; i < meshComp->GetNumMaterials(); i++) meshComp->SetMaterial(i, buyInfo.attachmentMesh->GetMaterial(i));
+		for(uint8 i = 0; i < meshComp->GetNumMaterials(); i++) meshComp->SetMaterial(i, buyInfo.attachmentMesh->GetMaterial(i));
 	}
 	// if(lockedMaterial) for(int i = 0; i < meshComp->GetNumMaterials(); i++) meshComp->SetMaterial(i, lockedMaterial);
 }
