+++ projProblox/Source/projProblox/Pickupables/Propeller.h
@@ -29,16 +29,16 @@ class PROJPROBLOX_API APropeller : public APickupableMaster
 	UBoxComponent* windBox;
 
 	UPROPERTY(EditDefaultsOnly, Category = "Ability|Forces", meta = (ToolTip = "The amount of force to apply to the thing that it's attached to"))
-	unsigned int propelForce = 50;
+	uint16 propelForce = 50;
 	
 	UPROPERTY(EditDefaultsOnly, Category = "Ability|Forces", meta = (Delta = 1, ToolTip = "The amount of force to apply to things that are inside the windCollider while active"))
 	float pushForce = 100;
 
 	UPROPERTY(EditDefaultsOnly, Category = "Ability|Aesthetic", meta = (Delta = 1, ToolTip = "The speed the propellers rotate (purely visual)."))
-	unsigned int spinSpeed = 22;
+	uint16 spinSpeed = 22;
 
 	UPROPERTY(EditDefaultsOnly, Category = "Ability|Forces", meta = (ToolTip = "How much the mass of the core affects the force of the propeller (higher number = higher force)."))
-	unsigned int massScale = 1000;
+	uint16 massScale = 1000;
 
 	
 protected:
